build: Revamp to use CMake
This will allow us to use the same build system on all platforms, and
will allow us to use the same build system for the Fortran code as well.

It will avoid the distutils removal issue on Python 3.12 and provide
better support for building on Windows.

In addition the phshift2007 programs are now built as part of the
package wheel build process, and are installed along with the package.